requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain ecosystem
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
langchain-anthropic==0.1.1

# Vector store and embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Document processing
pypdf==4.0.1
python-multipart==0.0.6

# UI
streamlit==1.31.0
plotly==5.18.0

# Utilities
numpy==1.26.3
pandas==2.2.0
tenacity==8.2.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0